Added files via upload
This will now run 99% of the time. There is an ugly hack that will allow for 2 clerks to run. the input is also now done via command line arguments instead of scanner. Lastly, I have added the use of the isAliveMethod so that the dragon and clerk will terminate after all the adventurers.