Added files via upload
These classes will implement the story, except for the use of the isInterrupted method as well as having the yield method make the adventurers not hog the dragon. Also for some reason the busy-wait for the shop will only fail if I don't put the print-line for checking the error in there.